// .repos/semantic-web/traqula/packages/core/lib/types.ts
export interface Typed<Type extends string = string> {
  type: Type;
  subType?: string;
}